Extract HTTP request to separate module
